sweet alert & blog page done

+++ src/Pages/MyToys/ToysMY.jsx
@@ -1,19 +1,40 @@
 import { Link } from 'react-router-dom';
 import { FaTrash } from "react-icons/fa";
+import Swal from 'sweetalert2';
 
 const ToysMY = ({ toy, index, toys, setToys }) => {
     const { _id, sellerName, toyName, price, category, quantity } = toy;
 
     const handleDelete = _id => {
-        fetch(`https://toy-marketplace-server-plum.vercel.app/toys/${_id}`, {
-            method: "DELETE"
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://toy-marketplace-server-plum.vercel.app/toys/${_id}`, {
+                    method: "DELETE"
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your file has been deleted.',
+                                'success'
+                            )
+                            const remaining = toys.filter(ty => ty._id !== _id);
+                            setToys(remaining);
+                        }
+                    })
+
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-            })
-        const remaining = toys.filter(ty => ty._id !== _id);
-        setToys(remaining);
+
     }
     return (
         <tr>
